Update SLF4J dependencies.
This commit brings the SLF4J dependency to the latest public release, while removing the JDK1.4 binding as an explicit dependency (it instead is made an optional dependency, which allows it to remain on the manifest classpath).

This change no longer introduces a specific binding when JSTUN is used as a dependency of another project. Such a project likely has its own binding preference.

As of SLF4J version 1.6, in the absence of a binding, SLF4J will default to a no-operation (NOP) logger implementation. A specific binding can still be provided, when executing JSTUN as a stand-alone application. To do so, simply place the binding JAR file next to the JSTUN JAR file.