It should be possible to stop the StunServer
The commit adds a flag that will allow the StunServer to gracefully shut down. Additionally, it introduces
a timeout that applies to data receival. When no data is received within the timeout, the newly added flag
is evaluated, after which the socket attempts to receive data again.

This is a fix for https://github.com/tking/JSTUN/issues/6